+++ src/test/java/racingcar/domain/numbergenerator/CarNameParserTest.java
@@ -1,12 +1,12 @@
 package racingcar.domain.numbergenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.CarName;
+import racingcar.domain.CarNameParser;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -15,7 +15,7 @@ class CarNameParserTest {
     void 이름_입력_파싱_테스트1() {
         String failInputNames = "dave";
 
-        assertThatThrownBy(()->CarNameParser.parse(failInputNames))
+        assertThatThrownBy(()-> CarNameParser.parse(failInputNames))
                 .isInstanceOf(IllegalArgumentException.class);
     }
